requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
python-telegram-bot==21.6
schedule==1.2.2
playwright==1.49.0
flask==3.1.0
